services/ai-py/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.82.0
anthropic==0.52.0
pydantic==2.11.3
pydantic-settings==2.9.1
pyyaml==6.0.2
httpx==0.28.1